由鹭青发起的VOD SDK自动发布, 版本号：2.10.0 发布日志： 1, Add the DescribePlayTopVideos  API,  to describe the top video statistics of play data. 2, Add the DescribePlayUserAvg API, to describe the average statistics of play data. 3, Add the DescribePlayUserTotal API, to describe the total statistics of play data. 4, Add the DescribePlayVideoStatis API, to describe the single video statistics of play data. 5, Add the DescribeRefreshQuota API, to get quota of refreshing cache. 6, Add the DescribeRefreshTasks API, to get status of refreshing cache. 7, Add the PushObjectCache API, to summit a job of pushing objects to cdn cache. 8, Add the RefreshObjectCaches API, to refresh cdn cache. 9, Add the SubmitTranscodeJobs API, to support the user to submit transocde job base with VideoId, TemplateGroupId, EncryptConfig and so on. 10, Add the SubmitSnapshotJob API, to support the user to submit snapshot job base with VideoId, SpecifiedOffsetTime, Width, Height and so on. 11, Update the ProduceEditingProjectVideo API, add MediaMetadata field of request to set the produced media metadata, add ProduceConfig field of request to set some customized config options, such as TemplateGroupId and so on.

+++ aliyun-net-sdk-vod/Vod/Transform/V20170321/GetVideoInfoResponseUnmarshaller.cs
@@ -45,7 +45,7 @@ namespace Aliyun.Acs.vod.Transform.V20170321
 			video.CreationTime = context.StringValue("GetVideoInfo.Video.CreationTime");
 			video.ModifyTime = context.StringValue("GetVideoInfo.Video.ModifyTime");
 			video.CoverURL = context.StringValue("GetVideoInfo.Video.CoverURL");
-			video.CateId = context.IntegerValue("GetVideoInfo.Video.CateId");
+			video.CateId = context.LongValue("GetVideoInfo.Video.CateId");
 			video.CateName = context.StringValue("GetVideoInfo.Video.CateName");
 
 			List<string> video_snapshots = new List<string>();
